Create proxy.js
adds smooth streaming and prefers manual thinking insertion over the prompt.

use a <STOP=(your-word-to-stop)> somewhere in your prompt to tell it where to insert the </think>. this is MUCH more consistent. works really well if you always make it start with something like an Info board, which i have left in by default.